Add a Python framework for PR handling
First iteration to create a Python framework for running tasks -
in the first instance the only task is PR fetching and testing.
Tasks are run in their own multiprocessing.Process and
messages may be passed between them using multiprocessing.Queues.
This structure is intended to improve on the current PR testing
by pushing the PR fetching into its own background task -
so that new PRs can be fetched asynchronously from testing
of already fetched ones - and by adding fetched PRs to a Priority
Queue - so that we may interrupt testing of previously seen PRs
if higher priority PRs i.e. new, are found.