Add timeout for git pull, modify cmds.Command
When handling a pull request, the initial git pull origin command
on the local repositories may hang so add a timeout to protect.
If timeout detected, the PR is dropped.

Removed the cmds.AliBuild/AliDoctor shortcuts, and instead
call the explicit cmds.Command with the executable name.
Adapted the cmds.Command constructor to take the string name
of the command instead of the sh.Command instance. The latter
is now created from the name inside the cmds.Command constructor.